chore: ajust of discossions missing
